refactoring order tests
