+++ src/test/java/user/BaseUserTest.java
@@ -1,0 +1,15 @@
+package user;
+
+import stellarburgres.user.User;
+import stellarburgres.user.UserClient;
+import stellarburgres.user.UserCredentials;
+
+public class BaseUserTest {
+
+    UserClient userClient = new UserClient();
+    User credentials;
+    User user;
+    UserCredentials userCredentials;
+    String token;
+
+}
